move editor frame into own file
