+++ .pylintrc
@@ -8,4 +8,5 @@ disable=
     import-error,
     no-else-return,
     broad-exception-caught,
-    too-few-public-methods
+    too-few-public-methods,
+    too-many-ancestors
