+++ cs2tracker/app/application.py
@@ -12,6 +12,7 @@ import sv_ttk
 from matplotlib.axes import Axes
 from matplotlib.dates import DateFormatter
 
+from cs2tracker.app.editor_frame import ConfigEditorFrame
 from cs2tracker.constants import (
     CONFIG_FILE,
     CONFIG_FILE_BACKUP,
@@ -32,9 +33,6 @@ WINDOW_SIZE = "630x380"
 CONFIG_EDITOR_TITLE = "Config Editor"
 CONFIG_EDITOR_SIZE = "800x750"
 
-NEW_CUSTOM_ITEM_TITLE = "Add Custom Item"
-NEW_CUSTOM_ITEM_SIZE = "500x200"
-
 SCRAPER_WINDOW_HEIGHT = 40
 SCRAPER_WINDOW_WIDTH = 120
 SCRAPER_WINDOW_BACKGROUND_COLOR = "Black"
@@ -44,7 +42,6 @@ class Application:
     def __init__(self):
         self.scraper = Scraper()
         self.application_window = None
-        self.config_editor_window = None
 
     def run(self):
         """Run the main application window with buttons for scraping prices, editing the
@@ -222,201 +219,14 @@ class Application:
             # TODO: implement external window for Linux
             self.scraper.scrape_prices()
 
-    def _make_tree_editable(self, editor_frame, tree):
-        """
-        Add a binding to the treeview that allows double-clicking on a cell to edit its
-        value.
-
-        Source: https://stackoverflow.com/questions/75787251/create-an-editable-tkinter-treeview-with-keyword-connection
-        """
-
-        def set_cell_value(event):
-            def save_edit(event):
-                tree.set(row, column=column, value=event.widget.get())
-                event.widget.destroy()
-
-            try:
-                row = tree.identify_row(event.y)
-                column = tree.identify_column(event.x)
-                item_text = tree.set(row, column)
-                if item_text.strip() == "":
-                    left_item_text = tree.item(row, "text")
-                    # Don't allow editing of section headers
-                    if any(left_item_text == section for section in self.scraper.config.sections()):
-                        return
-                x, y, w, h = tree.bbox(row, column)
-                entryedit = ttk.Entry(editor_frame)
-                entryedit.place(x=x, y=y, width=w, height=h + 3)  # type: ignore
-                entryedit.insert("end", item_text)
-                entryedit.bind("<Return>", save_edit)
-                entryedit.focus_set()
-                entryedit.grab_set()
-            except Exception:
-                pass
-
-        def destroy_entries(_):
-            """Destroy any entry widgets in the treeview when the mouse wheel is
-            used.
-            """
-            for widget in editor_frame.winfo_children():
-                if isinstance(widget, ttk.Entry):
-                    widget.destroy()
-
-        def destroy_entry(event):
-            """Destroy the entry widget if the user clicks outside of it."""
-            if isinstance(event.widget, ttk.Entry):
-                event.widget.destroy()
-
-        tree.bind("<Double-1>", set_cell_value)
-        self.config_editor_window.bind("<MouseWheel>", destroy_entries)  # type: ignore
-        self.config_editor_window.bind("<Button-1>", destroy_entry)  # type: ignore
-
-    def _configure_treeview(self, editor_frame):
-        """Add a treeview to the editor frame to display and edit configuration
-        options.
-        """
-        scrollbar = ttk.Scrollbar(editor_frame)
-        scrollbar.pack(side="right", fill="y", padx=(5, 0))
-
-        tree = ttk.Treeview(
-            editor_frame,
-            columns=(1,),
-            height=10,
-            selectmode="browse",
-            yscrollcommand=scrollbar.set,
-        )
-        scrollbar.config(command=tree.yview)
-
-        tree.column("#0", anchor="w", width=200)
-        tree.column(1, anchor="w", width=25)
-        tree.heading("#0", text="Option")
-        tree.heading(1, text="Value")
-
-        for section in self.scraper.config.sections():
-            if section == "App Settings":
-                continue
-            section_level = tree.insert("", "end", iid=section, text=section)
-            for config_option, value in self.scraper.config.items(section):
-                title_option = config_option.replace("_", " ").title()
-                tree.insert(section_level, "end", text=title_option, values=(value,))
-
-        self._make_tree_editable(editor_frame, tree)
-
-        return tree
-
-    def _configure_save_button(self, button_frame, tree):
-        """Save updated options and values from the treeview back to the config file."""
-
-        def save_config():
-            for child in tree.get_children():
-                for item in tree.get_children(child):
-                    title_option = tree.item(item, "text")
-                    config_option = title_option.lower().replace(" ", "_")
-                    value = tree.item(item, "values")[0]
-                    section = tree.parent(item)
-                    section_name = tree.item(section, "text")
-                    if section_name == "Custom Items":
-                        # custom items are already saved upon creation (Saving them again would result in duplicates)
-                        continue
-                    self.scraper.config.set(section_name, config_option, value)
-
-            self.scraper.config.write_to_file()
-            if self.scraper.config.valid:
-                messagebox.showinfo(
-                    "Config Saved", "The configuration has been saved successfully."
-                )
-            else:
-                messagebox.showerror(
-                    "Config Error",
-                    f"The configuration is invalid. ({self.scraper.config.last_error})",
-                )
-
-        save_button = ttk.Button(button_frame, text="Save", command=save_config)
-        save_button.pack(side="left", expand=True, padx=5)
-
-    def _configure_custom_item_button(self, button_frame, tree):
-        """Add a button that opens an entry dialog to add a custom item to the
-        configuration.
-        """
-
-        def add_custom_item(item_url, item_owned):
-            """Add a custom item to the configuration."""
-            if not item_url or not item_owned:
-                messagebox.showerror("Input Error", "All fields must be filled out.")
-                return
-
-            try:
-                if int(item_owned) < 0:
-                    raise ValueError("Owned count must be a non-negative integer.")
-            except ValueError as error:
-                messagebox.showerror("Input Error", f"Invalid owned count: {error}")
-                return
-
-            self.scraper.config.set("Custom Items", item_url, item_owned)
-            self.scraper.config.write_to_file()
-            if self.scraper.config.valid:
-                tree.insert("Custom Items", "end", text=item_url, values=(item_owned,))
-                messagebox.showinfo("Custom Item Added", "Custom item has been added successfully.")
-            else:
-                self.scraper.config.remove_option("Custom Items", item_url)
-                messagebox.showerror(
-                    "Config Error",
-                    f"The configuration is invalid. ({self.scraper.config.last_error})",
-                )
-
-        def open_custom_item_dialog():
-            """Open a dialog to enter custom item details."""
-            dialog = tk.Toplevel(self.config_editor_window)
-            dialog.title(NEW_CUSTOM_ITEM_TITLE)
-            dialog.geometry(NEW_CUSTOM_ITEM_SIZE)
-
-            dialog_frame = ttk.Frame(dialog, padding=10)
-            dialog_frame.pack(expand=True, fill="both")
-
-            ttk.Label(dialog_frame, text="Item URL:").pack(pady=5)
-            item_url_entry = ttk.Entry(dialog_frame)
-            item_url_entry.pack(fill="x", padx=10)
-
-            ttk.Label(dialog_frame, text="Owned Count:").pack(pady=5)
-            item_owned_entry = ttk.Entry(dialog_frame)
-            item_owned_entry.pack(fill="x", padx=10)
-
-            add_button = ttk.Button(
-                dialog_frame,
-                text="Add",
-                command=lambda: add_custom_item(item_url_entry.get(), item_owned_entry.get()),
-            )
-            add_button.pack(pady=10)
-
-        custom_item_button = ttk.Button(
-            button_frame, text="Add Custom Item", command=open_custom_item_dialog
-        )
-        custom_item_button.pack(side="left", expand=True, padx=5)
-
-    def _configure_editor_frame(self):
-        """Configure the main editor frame which displays the configuration options in a
-        structured way.
-        """
-        editor_frame = ttk.Frame(self.config_editor_window, padding=30)
-        editor_frame.pack(expand=True, fill="both")
-
-        tree = self._configure_treeview(editor_frame)
-        tree.pack(expand=True, fill="both")
-
-        button_frame = ttk.Frame(editor_frame, padding=10)
-
-        self._configure_save_button(button_frame, tree)
-        self._configure_custom_item_button(button_frame, tree)
-
-        button_frame.pack(side="bottom", padx=10, pady=(0, 10))
-
     def _edit_config(self):
         """Open a new window with a config editor GUI."""
-        self.config_editor_window = tk.Toplevel(self.application_window)
-        self.config_editor_window.geometry(CONFIG_EDITOR_SIZE)
-        self.config_editor_window.title(CONFIG_EDITOR_TITLE)
+        config_editor_window = tk.Toplevel(self.application_window)
+        config_editor_window.geometry(CONFIG_EDITOR_SIZE)
+        config_editor_window.title(CONFIG_EDITOR_TITLE)
 
-        self._configure_editor_frame()
+        editor_frame = ConfigEditorFrame(config_editor_window, self.scraper.config)
+        editor_frame.pack(expand=True, fill="both")
 
     def _reset_config(self):
         """Reset the configuration file to its default state."""
